Fix/Feat: 처음 렌더링될 때 총수입/총지출 렌더링 누락 오류 수정 & 월 이동에 맞춰서 값 변경 추가
- 차트 전체데이터 패치 -> 먼슬리로 변경
