+++ src/App.vue
@@ -11,11 +11,17 @@
 <script setup>
 import { computed } from 'vue';
 import Header from '@/components/Header.vue';
-import { useAccountListStore } from '@/stores/account.js';
+// import { useAccountListStore } from '@/stores/account.js';
+import { useMonthlyAccountStore } from './stores/monthlyAccount';
 // import Loading from '@/components/Loading.vue'
 
-const AccountListStore = useAccountListStore();
+// const AccountListStore = useAccountListStore();
+
 // const isLoading = computed(()=>todoListStore.isLoading);
-const fetchAccountList = AccountListStore.fetchAccountList;
-fetchAccountList();
+// const fetchAccountList = AccountListStore.fetchAccountList;
+// fetchAccountList();
+
+const monthlyAccountStore = useMonthlyAccountStore();
+const {fetchMonthlyAccountList} = monthlyAccountStore;
+fetchMonthlyAccountList(new Date().getMonth() + 1);
 </script>
